Toolbox error dialog now has version number in title.
Side effect: "java -jar iText.jar" is enough to get the version number.

If the toolbox jar is present, the GUI can display the versions numbers
using "Help->Versions".

If the toolbox is not there, the iText version is in the dialog title.

This may help with customer support queries...


git-svn-id: https://itext.svn.sourceforge.net/svnroot/itext/trunk@2966 d55c716e-6631-0410-aaf4-a39e642303ef