pattern for test cases was too restrictive, some tests were not run
Conversion of TestPdfCopyAndStamp to use JUnit 4 annotations instead of
inheriting from TestCase

Two tests in TestPdfCopyAndStamp are tagged with @Ignore
as they result in failure or error. It looks like they only work
on Windows anyway and with a GUI so not suitable for automatic
testing. Maybe I'll move them to the test.swing part later on


git-svn-id: https://itext.svn.sourceforge.net/svnroot/itext/trunk@3663 d55c716e-6631-0410-aaf4-a39e642303ef